runm object get like property-definition get
Updates the `runm object get` CLI command to work like the `runm
property-definition get` command. It now accepts either a UUID argument
or a --filter CLI option.

The object --filter CLI option has been changed to accept a uuid or name
key in the filter string instead of just "object".
